Wounds and Vigor sheet integration, and some cleanup.
Finished the W&V sheet rendering and rolled that out to the NPC sheet as well.

I also took the liberty of cleaning up the structure a little bit, and made the tooltips less 'RAWR!' by restricting them only to the header or value, instead of the entire section div.