Health Configuration menu, with more control.
Added the ability to set any numeric rate of health computation desired, and to enforce an arbitrary number of maximized levels.  Also provided access to control the rounding function and application.

Health configuration was taking up quite a bit of the settings for the system already, so I rolled all of this into a new menu for managing health.  Unfortunately, this means I had to change a fair bit of the localization, and as I speak neither French nor German, my only options were to rely on Google Translate or just purge it.  If I purge it, someone might fix it - if it put in a mangled translation it might stay forever.