Fix circle templates not working in cell centers
Measure templates, when pf1.measureStyle was true, would not place in
cell centers, or, when dragged to cell centers, would highlight
cells incorrectly.

This was due to a few things:
* The point distance was measuring to was changed, depending on if the
origin was or was not on a cell border. We don't want to do this, as the
cells themselves never change position. I changed this to always use the
center of the cells.
* There was offsetting for non-cones that, with the other changes made,
did not need to be there.
* The maths in measureDistance was slightly off. I changed it to use a
simpler formula which is consistently correct, as well as adjusting the
order of some Math.ceil and Math.abs operations.

With these issues addressed, placement and measuring of circular
templates works much better. Cone and ray templates still need to be
tested, as they will likely have had their behaviour impacted. Templates
and shapes for larger tokens should also be considered.

(this was rather a painful mathsy one to dive into for my first go at
truly contributing to Foundry. Ouch.)