Fix cones
Adjust origin-shifting for cones, only do so once as the origin does not
change. This means that cones will measure from the borders of cells, a
la https://www.d20pfsrd.com/magic/#Aiming_a_Spell

Adjust cone placement rules to work more intuitively - you can place a
cone wherever, including the center of squares, but it will still
correctly measure from the border of that square, in the direction the
cone is pointing.