Fix #211
This doesn't properly fix the issue but it works in the current version of the sheet. The current way we patch getTemplate to preventIndent doesn't patch nested partials. If \<textareas\> and \<pre\> start to be more common in the more complex versions of the sheet, we may need to patch Handlebars at a deeper level.