chatMessage changes
I have a module that would like to roll ability, skills and saving throws.  Problem is I don't want the request to automatically open a Chat Message.  I've attempted some interested code to work around it, but DnD5e has a handy option of requesting that a dice roll doesn't open a ChatMessage, and I'd very kindly like to add it to Pathfinder.