Fix reach highlights stacking when shift-clicking quick action
Without this, there is no single reference to the current highlight,
which makes it necessary to handle the rendering and clearing
in various locations, duplicating code.
More specifically, it was possible to stack more and more highlights
by triggering attacks from a token's quick action HUD while skipping
the dialog – as the `mouseleave` event was not triggered, but the
`mouseenter` one was, a new highlight is added with every attack.