Use descriptions according to item's state as canonical description
This changes what data is stored in a chat message's `content` field so
that the used description mirrors the item's current identified state.
To allow GMs the usual quick access to all info however, a description
block is added containing the item's identified and enriched description
as it was when the item was rolled.
With that, the data presented in the changelog should no longer get
changed after the fact, and exported messages accurately show a
canonical item state.