Fix `D20RollPF` not being serialized properly
Using `Roll#toJSON` is not sufficient to generate a data object suitable
for serialization; `JSON.stringify` is necessary to ensure that not
only the roll's `toJSON` is used, but also the `toJSON` for all other
classes (e.g. the roll's `terms`).

As the `messageData` object holds dot-notation keys, `mergeObject`
calls `expandObject`, which would result in the `rolls` array not
containing the actual `Roll` instance or an object created through
`toJSON`, but a simple object containing the roll's properties instead.
To avoid that, the `rolls` array has to be placed outside the
`mergeObject` object.

Fixes #1715