build(packs): extract diffs to templates for docs and components
This commit refactors the pack extraction tools into multiple modules,
and changes the template enforcing so that only diffs to the applicable
document template are stored – default values are thus pruned.
This diffing process is extended to include Actions and Changes.

The file names of extracted pack entries now conform to a format of
"<name>.<id>.json" instead of "<name>.json", allowing multiple entries
in a pack to share a name.
The `keepIds` option is removed, as IDs can no longer be reliably
tracked. In cases of possible accidental ID changes a warning is logged
instead.