fix: re-enable calling applyChange for some Changes
This allows calling applyChange for Changes that are not meant to be
applied in the normal data prep workflow, but have to suffer the
function call so that their effective operator/value can be determined.
Ideally, value determination and application should not happen in one
go, but a rework has been postponed to a later date.

Refs: 9e36be3dab8ff02a3f1dfdf3678237eca021e664
Refs: #1070